fix: move modal components (again)
