+++ packages/shared/src/components/modals/common/types.ts
@@ -69,7 +69,6 @@ export enum LazyModal {
   MobileSmartPrompts = 'mobileSmartPrompts',
   GiveAward = 'giveAward',
   ContentModal = 'contentModal',
-  MostVisitedSites = 'mostVisitedSites',
   CustomLinks = 'customLinks',
 }
